+++ src/triangle.rs
@@ -1,4 +1,4 @@
-use bevy::math::f64::DVec3 as Vec3;
+use bevy::prelude::Vec3;
 use std::ops::{Add, Mul};
 use Vec3 as Color;
 
@@ -8,7 +8,7 @@ pub trait ToTriangle {
 
 pub struct Triangle<T>
 where
-    T: Copy + Add<T, Output = T> + Mul<f64, Output = T>,
+    T: Copy + Add<T, Output = T> + Mul<f32, Output = T>,
 {
     points: [T; 3],
     colors: [Color; 3],
@@ -17,7 +17,7 @@ where
 
 impl<T> Triangle<T>
 where
-    T: Copy + Add<T, Output = T> + Mul<f64, Output = T>,
+    T: Copy + Add<T, Output = T> + Mul<f32, Output = T>,
 {
     pub fn new(p: [T; 3], c: [Color; 3], n: [T; 3]) -> Self {
         Triangle {
